requirements.txt
blurhash==1.1.4
boost==0.1
charset-normalizer==2.1.1
cmake==3.24.2
cycler==0.11.0
decorator==5.1.1
dlib==19.24.0
dotmap==1.3.30
fonttools==4.38.0
greenlet==2.0.0
idna==3.4
imageio==2.22.3
importlib-metadata==5.0.0
imutils==0.5.4
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
kornia==0.5.3
llvmlite==0.36.0
MarkupSafe==2.1.1
Mastodon.py==1.5.1
matplotlib==3.5.3
networkx==2.6.3
numba==0.53.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.5.2.54
packaging==21.3
pandas==1.3.5
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.6
PyWavelets==1.3.0
PyYAML==6.0
render==1.0.0
requests==2.28.1
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.7.3
seaborn==0.11.1
six==1.16.0
SQLAlchemy==1.4.42
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.0
typing_extensions==4.4.0
urllib3==1.26.12
zipp==3.10.0
